First draft of KSI framework using JDK component as guinea pig.
At this point, KSI is feature-complete, at least for installing one
component (the JDK) into a specified service.  Instructions for
standard usage are at the top of "ksi.sh".

Basic preconditions for using KSI on a particular host are:

* The creation of a service user and group, e.g. "kualiadm" and "kuali",
  with access to that account via sudoers.  All KSI invocations are
  meant to be done as the appropriate service user.

* The existence of the /optimized base directories for the service,
  e.g. /opt/kuali, /etc/opt/kuali, and /var/opt/kuali; these should
  be owned and grouped to the service user and group.

* The availability of the KSI installation repository, either as
  a git clone, or via some other mechanism.

Note that as of this commit, things have only been verified to
work on the Toolkit host, as a normal user, installing to a
fake /optimized directory structure under a home directory.  Until
we've done additional work on the DEV virtual host, YMMV.

Mike, 7/5/2012