Added RAMBLE.md document explaining concepts and flow.
This commit adds one file to the top of the project, "RAMBLE.md".
This document contains a lengthy (and informal) explanation of
the basic design concepts behind the KSI architecture (at this
point), and a walk-through of what happens when you actually
invoke "ksi.sh".  It is intended as a supplement to information
already available in comments in various *.sh files within the
project.