Third draft of KSI framework, including many tweaks from piloting in the DEV environment.
This commit adds the Oracle Instant Client component, and rewires
the way we're handling Tomcat installations, so that now you use
one "install" action (for the TOMCAT_HOME contents), and then
multiple "mkinst" actions (one for each TOMCAT_BASE that you want
to create).  It's still a little grubby, but it works better than
the previous version.

Additional tweaks include:

* The JTool third-party Java library now installs as part of the
  JDK component.

* The global "web.xml" configuration file for Tomcat now has
  trimSpaces turned on (recommended for bandwidth efficiency
  and general nice-to-be-able-to-read-the-page-source goodness.

* The options passed to the JVM during Tomcat startup now point
  to /dev/urandom instead of /dev/random for random number
  generation (preserving a fix just recently applied to the
  old-style PRD environment).

This will hopefully lock off the KSI installations of the JDK,
Tomcat, and Oracle Instance Client components for the current
round of environment /optimization -- I need to add the Ruby
interpreter (separate branch and commit forthcoming), and then
we should probably tag a "1.0" version or something similar.

Mike, 9/14/2012