First draft of adding kuali.conf generation to KSI.
This commit includes a first pass at generating kuali.conf in a
standardized fashion out of KSI; it provides a "mkconf" action
under an "apache" component.  Note that we're not compiling or
installing Apache at this point, just using the KSI service and
template functionality to standardize the contents of kuali.conf.

The generated file will wind up under /etc/opt/apache/apache-<version>,
from whence it can be installed to the appropriate path on the
webserver hosts.  Also note that the template file used to generate
this configuration is likely to change shortly, once we get the
SUP/PRD kuali.conf configuration stabilized as part of work on
KATTS-1093.